Refatored messd-up and missed-opportunities
