+++ include/interfaces.hpp
@@ -1,20 +1,10 @@
-#include <gpio.hpp>
-
+#ifndef __INTERFACES
+#define __INTERFACES
 using namespace std;
 
 class ModuleInterface {
 public:
   virtual void init();
   virtual void process(float msDelta);
-
-  virtual pin_t *getInputPinSchematic();
-  virtual pin_t *getOutputPinSchematic();
-
-  virtual double *getInputBuffer();
-  virtual double *getOutputBuffer();
-
-  virtual uint8_t getNumInputs();
-  virtual uint8_t getNumOutputs();
 };
-
-ModuleInterface *buildModule();
+#endif
