+++ modules/messd-up/messd-up.cpp
@@ -1,136 +1,113 @@
-#include "interfaces.hpp"
-#include <Arduino.h>
-#include <cutemodules.h>
+#include "messd-up.hpp"
 
-#define NUM_INPUTS 3
-#define NUM_OUTPUTS 4
+void Module::_scaleValues() {
+  this->ins.tempo = (this->ins.tempo > 0 ? this->ins.tempo : 1);
+  this->ins.beatsPerMeasure =
+      (this->ins.beatsPerMeasure > 0 ? this->ins.beatsPerMeasure : 1);
+  this->ins.subdivisionsPerMeasure =
+      (this->ins.subdivisionsPerMeasure > 0 ? this->ins.subdivisionsPerMeasure
+                                            : 1);
+  this->ins.phase =
+      (this->ins.phase <= 1 && ins.phase >= 0 ? this->ins.phase : 0);
+  this->ins.pulseWidth = (this->ins.pulseWidth < 1 && this->ins.pulseWidth > 0
+                              ? this->ins.pulseWidth
+                              : 0.5);
+}
 
-// Old arduino values
-// static pin_t inputPinSchematic[] = {
-//     {5, INPUT, DIGITAL}, // Mux 1 - A
-//     {6, INPUT, DIGITAL}, // Mux 1 - B
-//     {7, INPUT, DIGITAL}, // Mux 1 - C
-//     {A0, INPUT, ANALOG}, // Mux 1 - Read Pin
-// };
+void Module::_processEncoders() {
+  int inc =
+      encoder_process(&beat_encoder, digital_mux.outputs[DigitalMux.BEAT_ENC_A],
+                      digital_mux.outputs[DigitalMux.BEAT_ENC_B]);
+  if (inc != 0) {
+    state.beats += inc;
+    if (state.beats < beatsDivMin)
+      state.beats = beatsDivMax;
+    if (state.beats > beatsDivMax)
+      state.beats = beatsDivMin;
+  }
+  inc = encoder_process(&div_encoder,
+                        digital_mux.outputs[DigitalMux.DIVIDE_ENC_A],
+                        digital_mux.outputs[DigitalMux.DIVIDE_ENC_B]);
+  if (inc != 0) {
+    state.div += inc;
+    if (state.div < beatsDivMin)
+      state.div = beatsDivMax;
+    if (state.div > beatsDivMax)
+      state.div = beatsDivMin;
+  }
+}
 
-static pin_t inputPinSchematic[] = {
-    {PD4, INPUT, DIGITAL}, // Mux 1 - A
-    {PD3, INPUT, DIGITAL}, // Mux 1 - B
-    {PD2, INPUT, DIGITAL}, // Mux 1 - C
-    {PC0, INPUT, ANALOG}, // Mux 1 - Read Pin
-};
+Module::Module() { MS_init(&this->messd); };
 
-// Old arduino values
-static pin_t outputPinSchematic[] = {
-    {12, OUTPUT, DIGITAL}, // Clock Pin,
-    {11, OUTPUT, DIGITAL}, // Data Pin
-    {8, OUTPUT, DIGITAL},  // Latch Pin,
-};
+void Module::init() {
+  // Analog read pin for the digital mux
+  pinMode(digitalMuxIn, INPUT);
 
-// typedef enum muxInputNames1 {
-//   TRUNCATE_POT,
-//   DIVIDE_POT,
-//   TRUNCATE_CV,
-//   DIVIDE_CV,
-//   DIVIDE_SWITCH,
-//   BEAT_SWITCH,
-//   ROUND_SWITCH,
-//   LATCH_SWITCH
-// } inputNames_t;
+  // Setting all of the shift register pins to be outputs
+  pinMode(clockPinA, OUTPUT);
+  pinMode(latchPinA, OUTPUT);
+  pinMode(dataPinA, OUTPUT);
 
-typedef enum inputNames {
-  BEATS,
-  SUBDIVISIONS,
-  DEBUG_TEMPO
-} inputNames_t;
+  // Mux selector pins are outputs
+  pinMode(m0, OUTPUT);
+  pinMode(m1, OUTPUT);
+  pinMode(m2, OUTPUT);
+}
 
-typedef enum outputNames {
-  BEATS_OUT,
-  DOWNBEAT_OUT,
-  SUBDIVISIONS_OUT,
-  PHASE_OUT,
-} outputNames_t;
+void Module::process(float msDelta) {
+  _processEncoders();
+  mux_process(&digital_mux);
+  mux_process(&analog_mux);
 
-class Module : public ModuleInterface {
-private:
-  uint8_t numInputs;
-  uint8_t numOutputs;
-
-  double *inputBuffer;
-  double *outputBuffer;
-
-  messd_t messd;
-  messd_ins_t ins;
-  messd_outs_t outs;
-
-  void _scaleValues() {
-    this->ins.tempo = (this->ins.tempo > 0 ? this->ins.tempo : 1);
-    this->ins.beatsPerMeasure =
-        (this->ins.beatsPerMeasure > 0 ? this->ins.beatsPerMeasure : 1);
-    this->ins.subdivisionsPerMeasure =
-        (this->ins.subdivisionsPerMeasure > 0 ? this->ins.subdivisionsPerMeasure
-                                              : 1);
-    this->ins.phase =
-        (this->ins.phase <= 1 && ins.phase >= 0 ? this->ins.phase : 0);
-    this->ins.pulseWidth = (this->ins.pulseWidth < 1 && this->ins.pulseWidth > 0
-                                ? this->ins.pulseWidth
-                                : 0.5);
+  if (digitCounter == 4) {
+    digitCounter = 0;
   }
 
-public:
-  Module(uint8_t numInputs, uint8_t numOutputs) {
-    this->numInputs = numInputs;
-    this->numOutputs = numOutputs;
+  int value;
 
-    this->inputBuffer = (double *)malloc(sizeof(double) * numInputs);
-    this->outputBuffer = (double *)malloc(sizeof(double) * numOutputs);
-  };
+  switch (digitCounter) {
+  case 0:
+    value = state.div / 10;
+    break;
+  case 1:
+    value = state.div % 10;
+    break;
+  case 2:
+    value = state.beats / 10;
+    break;
+  case 3:
+    value = state.beats % 10;
+    break;
+  }
+  seven_segment_process(&shift_register, digitCounter, value);
+  digitCounter++;
+  /* this->ins.delta = msDelta / 1000.0; */
+  /* this->ins.tempo = this->inputBuffer[TEMPO]; */
+  /* this->ins.tempo = this->inputBuffer[DEBUG_TEMPO]; */
+  /* this->ins.beatsPerMeasure = this->inputBuffer[BEATS]; */
+  /* this->ins.subdivisionsPerMeasure = this->inputBuffer[SUBDIVISIONS]; */
+  // this->ins.phase = this->inputBuffer[PHASE];
+  // this->ins.beatsPerMeasure = 4;
+  // this->ins.subdivisionsPerMeasure = 7;
+  /* this->ins.phase = 0; */
 
-  void init() { MS_init(&this->messd); };
+  /* this->ins.ext_clock = 0; */
+  /* this->ins.truncation = 0; */
+  /* this->ins.metricModulation = 0; */
+  /* this->ins.latchChangesToDownbeat = 0; */
+  /* this->ins.invert = 0; */
+  /* this->ins.isRoundTrip = 0; */
+  /* this->ins.reset = 0; */
 
-  void process(float msDelta) {
-    this->ins.delta = msDelta / 1000.0;
-    /* this->ins.tempo = this->inputBuffer[TEMPO]; */
-    this->ins.tempo = this->inputBuffer[DEBUG_TEMPO];
-    this->ins.beatsPerMeasure = this->inputBuffer[BEATS];
-    this->ins.subdivisionsPerMeasure = this->inputBuffer[SUBDIVISIONS];
-    // this->ins.phase = this->inputBuffer[PHASE];
-    // this->ins.beatsPerMeasure = 4;
-    // this->ins.subdivisionsPerMeasure = 7;
-    this->ins.phase = 0;
+  /* this->ins.wrap = 0; */
+  /* this->ins.pulseWidth = 0.5; */
 
-    this->ins.ext_clock = 0;
-    this->ins.truncation = 0;
-    this->ins.metricModulation = 0;
-    this->ins.latchChangesToDownbeat = 0;
-    this->ins.invert = 0;
-    this->ins.isRoundTrip = 0;
-    this->ins.reset = 0;
+  /* _scaleValues(); */
 
-    this->ins.wrap = 0;
-    this->ins.pulseWidth = 0.5;
+  /* MS_process(&this->messd, &this->ins, &this->outs); */
 
-    _scaleValues();
-
-    MS_process(&this->messd, &this->ins, &this->outs);
-
-    this->outputBuffer[BEATS_OUT] = this->outs.beat;
-    this->outputBuffer[SUBDIVISIONS_OUT] = this->outs.subdivision;
-    this->outputBuffer[DOWNBEAT_OUT] = this->outs.downbeat;
-    this->outputBuffer[PHASE_OUT] = this->outs.phase;
-  };
-
-  pin_t *getInputPinSchematic() { return inputPinSchematic; };
-
-  pin_t *getOutputPinSchematic() { return outputPinSchematic; };
-
-  double *getInputBuffer() { return this->inputBuffer; };
-
-  double *getOutputBuffer() { return this->outputBuffer; };
-
-  uint8_t getNumInputs() { return this->numInputs; };
-
-  uint8_t getNumOutputs() { return this->numOutputs; };
+  /* this->outputBuffer[BEATS_OUT] = this->outs.beat; */
+  /* this->outputBuffer[SUBDIVISIONS_OUT] = this->outs.subdivision; */
+  /* this->outputBuffer[DOWNBEAT_OUT] = this->outs.downbeat; */
+  /* this->outputBuffer[PHASE_OUT] = this->outs.phase; */
 };
-
-ModuleInterface *buildModule() { return new Module(NUM_INPUTS, NUM_OUTPUTS); }
