+++ platformio.ini
@@ -38,11 +38,18 @@ extends = env:uno
 lib_deps =
   ${env:uno.lib_deps}
   modules/messd-up
+build_flags =
+  ${env:uno.build_flags}
+  -DMESSD_UP
 
 [env:missed-opportunities]
 extends = env:uno
 lib_deps =
+  ${env:uno.lib_deps}
   modules/missed-opportunities
+build_flags =
+  ${env:uno.build_flags}
+  -DMISSED_OPPORTUNITIES
 
 ; ==============================================
